Gra/requirements.txt
Pillow==11.2.1
pygame==2.6.1
